+++ R/defaults.R
@@ -37,7 +37,7 @@ default_toggles <- function() {
   list(cloglog = 0,
        movement = 0,
        est_surv = 0,
-       est_init = 1,
+       est_init = 0,
        ar_re = 0,
        iid_re = 0,
        sp_re = 0)